Seed the statutory rates the mockups quote
Overtime brackets, the employee-side BPJS rates, Bekasi's minimum wage, ten
of the fourteen quality-gate rules, and the credit scoring weights. Each
carries the citation the mockup prints beside it.

Seeded only what a mockup states on screen. The employer shares of BPJS, the
JP wage cap, and the whole TER tax table are real obligations that nothing in
the mockups quotes, so they are absent with a note rather than filled in from
memory. G2 will refuse to compute tax until the PMK 168/2023 schedule is
loaded, which is the right failure: a plausible wrong rate ships 412 payslips
before anyone notices, and an empty table stops at the first.

The holiday-within-8h band is a range in the source material (2.0-3.0x), so
it is stored as floor and ceiling with the rule that picks between them still
open, rather than quietly resolving to the floor.

Verified against PostgreSQL 16: all ten migrations apply clean, and the
scoring weights reproduce the mockup's worked example exactly - normalised
92/88/84/78 gives 152/121/116/86 and a total of 775 out of 850. That
round-trip is why max_points became numeric: the weights split 550 into
165/137.5/137.5/110, and whole numbers put the maximum at 851.

+++ supabase/migrations/20260922000008_credit_and_lending.sql
@@ -15,7 +15,10 @@ create table credit_score_factors (
   code varchar(40) not null unique,
   name varchar(255) not null,
   weight numeric(5,2) not null,                  -- 30 = 30%
-  max_points integer not null,
+  -- Numeric, not integer: the weights divide the 550 points above the base
+  -- into 165 / 137.5 / 137.5 / 110, and rounding those to whole numbers puts
+  -- the maximum at 851 instead of 850.
+  max_points numeric(8,2) not null,
   source varchar(100),                           -- 근태마감(G1) / 대출상환원장 / 인사마스터
   active boolean not null default true,
   effective_from date not null default current_date,
